refactored indeed-scraper to nest-puppeteer module
