+++ package.json
@@ -29,6 +29,7 @@
     "@nestjs/swagger": "^5.0.9",
     "@nestjs/typeorm": "^8.0.2",
     "bluebird": "^3.7.2",
+    "nest-puppeteer": "^1.1.1",
     "pg": "^8.7.1",
     "puppeteer": "^10.2.0",
     "reflect-metadata": "^0.1.13",
