+++ src/jobs/jobs.module.ts
@@ -4,10 +4,14 @@ import { JobsController } from './jobs.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Job } from './entities/job.entity';
 import { Company } from './entities/company.entity';
+import { PuppeteerModule } from 'nest-puppeteer';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Job, Company])],
+  imports: [
+    TypeOrmModule.forFeature([Job, Company]),
+    PuppeteerModule.forRoot(),
+  ],
   controllers: [JobsController],
-  providers: [JobsService]
+  providers: [JobsService],
 })
 export class JobsModule {}
